+++ tools/migration/changelog.md
@@ -54,3 +54,9 @@ Changelog for harbor database schema
 
   - create table `project_metadata`
   - insert data into table `project_metadata`
+
+## 1.4.0
+  - create table `replication_filter`
+  - create table `replication_filter_type`
+  - insert data into table `replication_filter_type`
+